requirements.txt
discord.py
lupa